requirements.txt
# background jobs
celery==4.3.0
redis==3.2.1

# Django 
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
drf-extensions==0.4.0
Markdown==3.1
requests==2.21.0

# Google cloud storage servics
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
gunicorn==19.9.0

# opencv
opencv-contrib-python==4.1.0.25
Pillow==6.0.0

# postgres driver
psycopg2-binary==2.8.1

